Remove vectorised numpy operations

+++ src/shapes/checkered_sphere.py
@@ -2,6 +2,6 @@ from src.shapes.sphere import Sphere
 
 class CheckeredSphere(Sphere):
     def diffuseColourAt(self, M):
-        checker = ((M.x * 2).astype(int) % 2) == ((M.z * 2).astype(int) % 2)
+        checker = int(M.x * 2) % 2 == int(M.z * 2) % 2
         checker ^= (M.x < 0) ^ (M.z < 0)
         return self.diffuse_colour * checker
